% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{DGP}
\alias{DGP}
\alias{ddgp}
\alias{pdgp}
\alias{qdgp}
\alias{rdgp}
\title{Discrete Generalized Pareto Distribution}
\usage{
ddgp(x, mu = 0, scale = 1, shape = 1)

pdgp(q, mu = 0, scale = 1, shape = 1, lower.tail = TRUE, log.p = FALSE)

qdgp(p, mu = 0, scale = 1, shape = 1, lower.tail = TRUE, log.p = FALSE)

rdgp(n, mu = 0, scale = 1, shape = 1)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{mu}{location parameter}

\item{scale}{scale parameter}

\item{shape}{shape parameter}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X \%leq x]}, otherwise,
\code{P[X > x]}.}

\item{p}{vector of probabilities}

\item{n}{number of observations}

\item{log, log.p}{logical; if TRUE, probabilities \code{p} are given as \code{log(p)}.}
}
\value{
\code{ddgp} gives the density, \code{pdgp} gives the distribution function,
\code{qdgp} gives the quantile function, and \code{rdgp} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the discrete
generalized Pareto distribution with parameters \code{mu}, \code{scale}, and \code{shape}.
}
