% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGPD.R
\name{DGPD}
\alias{DGPD}
\alias{ddgpd}
\title{Discrete Generalized Pareto Distribution}
\usage{
ddgpd(x, alpha = 1, lambda = 1, mu = 0, log = FALSE)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{alpha}{shape parameter}

\item{lambda}{scale parameter}

\item{mu}{location parameter}

\item{log, log.p}{logical; if TRUE, probabilities \code{p} are given as \code{log(p)}.}

\item{p}{vector of probabilities}

\item{n}{number of observations}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X \leq x]}, otherwise,
\code{P[X > x]}.}
}
\value{
\code{ddgpd} gives the density, \code{pdgpd} gives the distribution function,
\code{qdgpd} gives the quantile function, and \code{rdgpd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the discrete
generalized Pareto distribution with parameters \code{alpha}, \code{lambda}, and \code{mu}.
}
