% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHNBINOM.R
\name{DHNBINOM}
\alias{DHNBINOM}
\alias{ddhnbinom}
\title{Double-Hurdle Model for Negative Binomial Distribution}
\usage{
ddhnbinom(
  x,
  theta = 0.5,
  theta_1 = 0.1,
  size = 0.9,
  mu = 0.9,
  scale = 1,
  shape = 1,
  threshold = 3,
  log = FALSE
)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{theta}{zero generating parameter (probability of zeros)}

\item{theta_1}{probability of a count at or above \code{mu}, conditional on non-zero count}

\item{size}{number of successes (dispersion parameter)}

\item{threshold}{second hurdle}

\item{log, log.p}{logical; if TRUE, probabilities \code{p} are given as \code{log(p)}.}

\item{p}{vector of probabilities}

\item{n}{number of observations}

\item{prob}{probability of success in each trial. 0 < prob < 1}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X \leq x]}, otherwise,
\code{P[X > x]}.}
}
\value{
\code{dhnbinom} gives the density, \code{phnbinom} gives the distribution function,
\code{qhnbinom} gives the quantile function, and \code{rhnbinom} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the hurdle model
using negative binomial distribution with parameters \code{theta}, \code{size}, and \code{prob}.
}
