% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHPOIS.R
\name{DHPOIS}
\alias{DHPOIS}
\alias{doubletruc}
\title{Double-Hurdle Poisson}
\usage{
doubletruc()
}
\arguments{
\item{x, q}{vector of quantiles}

\item{p}{vector of probabilities}

\item{n}{number of observations}

\item{theta}{zero-generating parameter (probability of zeros)}

\item{lambda}{expected Poisson count}

\item{theta_1}{}

\item{mu}{location parameter}

\item{scale}{scale parameter}

\item{shape}{shape parameter}

\item{log, log.p}{logical; if TRUE, probabilities \code{p} are given as \code{log(p)}.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X \leq x]}, otherwise,
\code{P[X > x]}.}
}
\value{
\code{ddhpoisdgp} gives the density, \code{pdhpoisdgp} gives the distribution function,
\code{qdhpoisdgp} gives the quantile function, and \code{rdhpoisdgp} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the double-hurdle
Poisson distribution with parameters \code{theta}, \code{lambda}, \code{theta_1}, \code{mu},
\code{scale}, \code{shape}, and \code{threshold}.
}
