% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLNORM.R
\name{DLNORM}
\alias{DLNORM}
\alias{ddlnorm}
\title{Discrete Lognormal Distribution}
\usage{
ddlnorm(x, meanlog = 0, sdlog = 1)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{meanlog}{mean on the log scale}

\item{sdlog}{stanard deviation on the log scale}

\item{p}{vector of probabilities}

\item{n}{number of observations}

\item{log, log.p}{logical; if TRUE, probabilities \code{p} are given as \code{log(p)}.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X \leq x]}, otherwise,
\code{P[X > x]}.}
}
\value{
\code{ddlnorm} gives the density, \code{pdlnorm} gives the distribution function,
\code{qdlnorm} gives the quantile function, and \code{rdlnorm} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the discrete
generalized Pareto distribution with parameters \code{meanlog} and \code{sdlog}.
}
