% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGPD.R
\name{HGPD}
\alias{HGPD}
\alias{dhgpd}
\title{Hurdle Model for Generalized Pareto Distribution}
\usage{
dhgpd(x, theta, mu = 1, sigma = 1, xi = 1, log = FALSE)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{theta}{zero generating parameter (probability of zeros)}

\item{mu}{location parameter}

\item{sigma}{(non-negative) scale parameter}

\item{xi}{shape parameter}

\item{log, log.p}{logical; if TRUE, probabilities \code{p} are given as \code{log(p)}.}

\item{p}{vector of probabilities}

\item{n}{number of observations}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X \leq x]}, otherwise,
\code{P[X > x]}.}
}
\value{
\code{dhgpd} gives the density, \code{phgpd} gives the distribution function,
\code{qhgpd} gives the quantile function, and \code{rhgpd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the hurdle model of the
generalized Pareto distribution with parameters \code{theta}, \code{mu}, \code{sigma}, and \code{xi}.
}
