% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLNORM.R
\name{HLNORM}
\alias{HLNORM}
\alias{dhlnorm}
\title{Hurdle Model for Log-normal Distribution}
\usage{
dhlnorm(x, theta, mu = 1, sigma = 1, xi = 1)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{theta}{zero generating parameter (probability of zeros)}

\item{p}{vector of probabilities}

\item{n}{number of observations}

\item{meanlog}{mean (log scale)}

\item{sdlog}{standard deviation (log scale)}

\item{log, log.p}{logical; if TRUE, probabilities \code{p} are given as \code{log(p)}.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X ??? x]}, otherwise,
\code{P[X > x]}.}
}
\value{
\code{dhlnorm} gives the density, \code{phlnorm} gives the distribution function,
\code{qhlnorm} gives the quantile function, and \code{rhlnorm} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the hurdle model of
log-normal distribution with parameters \code{theta}, \code{meanlog}, and \code{sdlog}.
}
