% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HNBINOM.R
\name{HNBINOM}
\alias{HNBINOM}
\alias{dhnbinom}
\title{Hurdle Model for Negative Binomial Distribution}
\usage{
dhnbinom(x, theta = 0.5, size = 1, mu = 1, log = FALSE)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{theta}{zero generating parameter (probability of zeros)}

\item{size}{number of successes (dispersion parameter)}

\item{mu}{mean parameter (alternative parameterization)}

\item{log, log.p}{logical; if TRUE, probabilities \code{p} are given as \code{log(p)}.}

\item{p}{vector of probabilities}

\item{n}{number of observations}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X \leq x]}, otherwise,
\code{P[X > x]}.}
}
\value{
\code{dhnbinom} gives the density, \code{phnbinom} gives the distribution function,
\code{qhnbinom} gives the quantile function, and \code{rhnbinom} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the hurdle model
using negative binomial distribution with parameters \code{theta}, \code{size}, and \code{mu}.
}
