% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPOIS.R
\name{HPOIS}
\alias{HPOIS}
\alias{dhpois}
\title{Hurdle Model for Poisson Distribution}
\usage{
dhpois(x, theta = 0.5, lambda = 1, log = FALSE)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{theta}{zero-generating parameter (probability of zeros)}

\item{log, log.p}{logical; if TRUE, probabilities \code{p} are given as \code{log(p)}.}

\item{p}{vector of probabilities}

\item{n}{number of observations}

\item{lamda}{expected Poisson count}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X \leq x]}, otherwise,
\code{P[X > x]}.}
}
\value{
\code{dhpois} gives the density, \code{phpois} gives the distribution function,
\code{qhpois} gives the quantile function, and \code{rhpois} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the hurdle model
using Poisson distribution with parameters \code{theta} and \code{lambda}.
}
