% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIGPD.R
\name{ZIGPD}
\alias{ZIGPD}
\alias{dzigpd}
\title{Zero-inflated Generalized Pareto Distribution}
\usage{
dzigpd(x, theta, mu = 1, sigma = 1, xi = 1, log = FALSE)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{theta}{zero-inflation parameter (probability of zeros)}

\item{mu}{location parameter}

\item{sigma}{(non-negative) scale parameter}

\item{xi}{shape parameter}

\item{log, log.p}{logical; if TRUE, probabilities \code{p} are given as \code{log(p)}.}

\item{p}{vector of probabilities}

\item{n}{number of observations}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X \leq x]}, otherwise,
\code{P[X > x]}.}
}
\value{
\code{dzigpd} gives the density, \code{pzigpd} gives the distribution function,
\code{qzigpd} gives the quantile function, and \code{rzigpd} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the zero-inflated
generalized Pareto distribution with parameters \code{theta}, \code{mu}, \code{sigma}, and \code{xi}.
}
