% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHPOIS.R
\name{DHPOIS}
\alias{DHPOIS}
\alias{ddhpoisdgp}
\alias{pdhpoisdgp}
\alias{rdhpoisdgp}
\title{Double-Hurdle Poisson}
\usage{
ddhpoisdgp(
  x,
  theta = 0.5,
  threshold = 3,
  lambda = 1,
  theta_1 = 0.1,
  scale = 1,
  shape = 1
)

pdhpoisdgp(
  q,
  theta = 0.5,
  lambda = 1,
  threshold = 3,
  theta_1 = 0.1,
  scale = 1,
  shape = 1,
  lower.tail = TRUE,
  log.p = FALSE
)

rdhpoisdgp(
  n,
  theta = 0.5,
  lambda = 1,
  theta_1 = 0.1,
  scale = 1,
  shape = 1,
  threshold = 3
)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{theta}{zero-generating parameter (probability of zeros)}

\item{threshold}{second hurdle}

\item{lambda}{expected Poisson count}

\item{theta_1}{probability of a count at or above \code{mu}, conditional on non-zero count}

\item{scale}{scale parameter}

\item{shape}{shape parameter}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X \%leq x]}, otherwise,
\code{P[X > x]}.}

\item{n}{number of observations}

\item{p}{vector of probabilities}

\item{log, log.p}{logical; if TRUE, probabilities \code{p} are given as \code{log(p)}.}
}
\value{
\code{ddhpoisdgp} gives the density, \code{pdhpoisdgp} gives the distribution function,
\code{qdhpoisdgp} gives the quantile function, and \code{rdhpoisdgp} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the double-hurdle
Poisson distribution with parameters \code{theta}, \code{lambda}, \code{theta_1}, \code{mu},
\code{scale}, \code{shape}, and \code{threshold}.
}
