% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDGP.R
\name{HDGP}
\alias{HDGP}
\alias{dhdgp}
\title{Hurdle Model for Generalized Pareto Distribution}
\usage{
dhdgp(x, theta = 0.5, mu = 0, scale = 1, shape = 1, log = FALSE)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{theta}{zero generating parameter (probability of zeros)}

\item{mu}{location parameter}

\item{scale}{(non-negative) scale parameter}

\item{shape}{shape parameter}

\item{log, log.p}{logical; if TRUE, probabilities \code{p} are given as \code{log(p)}.}

\item{p}{vector of probabilities}

\item{n}{number of observations}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X \leq x]}, otherwise,
\code{P[X > x]}.}
}
\value{
\code{dhdgp} gives the density, \code{phdgp} gives the distribution function,
\code{qhdgp} gives the quantile function, and \code{rhdgp} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the hurdle model of the
generalized Pareto distribution with parameters \code{theta}, \code{mu}, \code{scale}, and \code{shape}.
}
