% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZINB.R
\name{ZINB}
\alias{ZINB}
\alias{dzinb}
\alias{pzinb}
\alias{qzinb}
\alias{rzinb}
\title{Zero-inflated Negative Binomial Distribution}
\usage{
dzinb(x, theta = 0.5, size = 1, mu = 1, log = FALSE)

pzinb(q, theta = 0.5, size = 1, mu = 1, lower.tail = TRUE, log.p = FALSE)

qzinb(p, theta = 0.5, size = 1, mu = 1, lower.tail = TRUE, log.p = FALSE)

rzinb(n, theta = 0.5, size = 1, mu = 1)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{theta}{zero-inflation parameter (probability of zeros)}

\item{size}{number of successes (dispersion parameter)}

\item{mu}{mean parameter (alternative parameterization)}

\item{log, log.p}{logical; if TRUE, probabilities \code{p} are given as \code{log(p)}.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X \%leq x]}, otherwise,
\code{P[X > x]}.}

\item{p}{vector of probabilities}

\item{n}{number of observations}
}
\value{
\code{dzinb} gives the density, \code{pzinb} gives the distribution function,
\code{qzinb} gives the quantile function, and \code{rzinb} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the zero-inflated
negative binomial distribution with parameters \code{theta}, \code{size}, and \code{mu}.
}
