% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIP.R
\name{ZIP}
\alias{ZIP}
\alias{dzip}
\alias{pzip}
\alias{qzip}
\alias{rzip}
\title{Zero-inflated Poisson Distribution}
\usage{
dzip(x, theta = 0.5, lambda = 1, log = FALSE)

pzip(q, theta = 0.5, lambda = 1, lower.tail = TRUE, log.p = FALSE)

qzip(p, theta = 0.5, lambda = 1, lower.tail = TRUE, log.p = FALSE)

rzip(n, theta = 0.5, lambda = 1)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{theta}{zero-inflation parameter (probability of zeros)}

\item{lambda}{expected Poisson count}

\item{log, log.p}{logical; if TRUE, probabilities \code{p} are given as \code{log(p)}.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X \%leq x]}, otherwise,
\code{P[X > x]}.}

\item{p}{vector of probabilities}

\item{n}{number of observations}
}
\value{
\code{dzip} gives the density, \code{pzip} gives the distribution function,
\code{qzip} gives the quantile function, and \code{rzip} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for the zero-inflated Poisson
distribution with parameters \code{theta} and \code{lambda}.
}
